Handle user not found case with nonzero exit code and stderr
If a user doesn't exist in the sharder db, report it as an error. This commit
doesn't change the core functionality of the sharder it is only used in the
(read only) admin find-user action. Doing this allows us to propagate a more
meaningful error admins trying to locate a user.

Signed-off-by: Ian Allison <iana@pims.math.ca>